requirements.txt
aniso8601==8.0.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.6.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
debugpy==1.6.2
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.5
entrypoints==0.4
executing==0.9.1
filelock==3.7.1
firebase==3.0.1
firebase-admin==5.2.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==1.0.1
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-dialogflow-cx==1.13.1
google-cloud-firestore==2.6.0
google-cloud-storage==2.4.0
google-cloud-vision==3.0.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.2
jsonify==0.5
jsonschema==4.9.1
jupyter-client==7.3.4
jupyter-core==4.11.1
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
msgpack==1.0.4
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
proto-plus==1.22.0
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==0.9.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2020.1
pywin32==304
pyzmq==23.2.0
requests==2.28.1
rsa==4.9
six==1.15.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.18
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
virtualenv==20.16.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
